Release v17.11-rc9
 Release v17.11-rc9  (#51)

+++ lnmp-docker.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
+# git remote add origin git@github.com:khs1994-docker/lnmp.git
+# git remote add aliyun git@code.aliyun.com:khs1994-docker/lnmp.git
+# git remote add tgit git@git.qcloud.com:khs1994-docker/lnmp.git
+# git remote add coding git@git.coding.net:khs1994/lnmp.git
 
 if [ "$1" = "development" -o "$1" = "production" ];then APP_ENV=$1; fi
 
@@ -109,6 +113,8 @@ gitbook(){
     -v $PWD/docs:/srv/gitbook-src \
     khs1994/gitbook \
     server
+
+  exit 0
 }
 
 dockerfile-update-sed(){
@@ -174,15 +180,17 @@ install_docker_compose(){
   command -v docker-compose >/dev/null 2>&1
   if [ $? = 0 ];then
     # 存在
-    DOCKER_COMPOSE_VERSION_CONTENT=`docker-compose --version`
-    if [ "$DOCKER_COMPOSE_VERSION_CONTENT" != "$DOCKER_COMPOSE_VERSION_CORRECT_CONTENT" ];then
-      print_error "`docker-compose --version` NOT installed Correct version, reinstall..."
-      sudo rm -rf `which docker-compose`
-      install_docker_compose
-      i=$(($i+1))
-      if [ $i -eq 2];then exit 1; fi
-    else
-      print_info "`docker-compose --version` already installed Correct version"
+    if [ ${OS} = Linux ];then
+      DOCKER_COMPOSE_VERSION_CONTENT=`docker-compose --version`
+      if [ "$DOCKER_COMPOSE_VERSION_CONTENT" != "$DOCKER_COMPOSE_VERSION_CORRECT_CONTENT" ];then
+        print_error "`docker-compose --version` NOT installed Correct version, reinstall..."
+        sudo rm -rf `which docker-compose`
+        install_docker_compose
+        i=$(($i+1))
+        if [ "$i" -eq 2 ];then exit 1; fi
+      else
+        print_info "`docker-compose --version` already installed Correct version"
+      fi
     fi
   else
     # 不存在
@@ -276,7 +284,7 @@ restore(){
 
 update(){
   GIT_STATUS=`git status -s`
-  if [ ! -z GIT_STATUS ];then print_error "Please commit then update"; exit 1; fi
+  if [ ! -z "${GIT_STATUS}" ];then git status -s; echo; print_error "Please commit then update"; exit 1; fi
   git fetch origin
   print_info "Branch is ${BRANCH}\n"
   if [ ${BRANCH} = "dev" ];then
